+++ src/thread_utils/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 '''
-Copyright 2014 Yoshida Shin
+Copyright 2014, 2015 Yoshida Shin
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@ limitations under the License.
 '''
 
 
-from error import Error, TimeoutError, DeadPoolError
+from error import Error, TimeoutError, DeadPoolError, CancelError
 from synchronized import synchronized
 from async import async, actor
 from pool import Pool